GRF-783 rename folder containing service
