+++ src/Akeneo/Category/back/Infrastructure/Symfony/Resources/config/cleaners.yml
@@ -3,4 +3,4 @@ services:
         arguments:
             - '@Akeneo\Category\Application\Storage\UpdateCategoryEnrichedValues'
 
-    Akeneo\Category\Domain\Query\DeactivatedTemplateAttributes\DeactivatedTemplateAttributesInValueCollectionFilter: ~
+    Akeneo\Category\Domain\Query\GetDeactivatedTemplateAttributes\DeactivatedTemplateAttributesInValueCollectionFilter: ~
